+++ tests/ValidationTest.php
@@ -2,22 +2,26 @@
 
 namespace Somnambulist\Components\Validation\Tests;
 
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use Somnambulist\Components\Validation\Factory;
+use Somnambulist\Components\Validation\Rule;
 use Somnambulist\Components\Validation\Validation;
+use function count;
+use function is_array;
 
 class ValidationTest extends TestCase
 {
     /**
      * @param string $rules
-     * @param array $expectedResult
+     * @param array  $expectedResult
      *
      * @dataProvider parseRuleProvider
      */
     public function testParseRule($rules, $expectedResult)
     {
-        $class = new ReflectionClass(Validation::class);
+        $class  = new ReflectionClass(Validation::class);
         $method = $class->getMethod('parseRule');
         $method->setAccessible(true);
 
@@ -65,6 +69,59 @@ class ValidationTest extends TestCase
                     ['/^([a-zA-Z\,]*)$/'],
                 ],
             ],
+            [
+                'date:Y-m-d H:i:s',
+                [
+                    'date',
+                    ['Y-m-d H:i:s'],
+                ],
+            ],
         ];
+    }
+
+    public function testValidateHandlesColonsInRuleAttributes()
+    {
+        $factory = new Factory();
+        $factory->addRule('order', new class extends Rule {
+            public function check(mixed $value): bool
+            {
+                return true;
+            }
+
+            public function fillParameters(array $params): self
+            {
+                Assert::assertEquals('id:ASC', $params[0]);
+                Assert::assertEquals('createdAt:DESC', $params[1]);
+
+                return $this;
+            }
+        });
+
+        $validation = $factory->validate([
+            'orderBy' => 'id:ASC',
+            'date'    => '2022-09-11 19:55:00',
+        ], [
+            'orderBy' => 'order:id:ASC,createdAt:DESC',
+            'date'    => [
+                'date:Y-m-d H:i:s',
+                'date' => 'Y-m-d H:i:s',
+            ]
+        ]);
+
+        $this->assertTrue($validation->passes());
+
+        $validation = $factory->validate([
+            'orderBy' => 'id:ASC',
+            'date'    => '2022-09-11',
+        ], [
+            'orderBy' => 'order:id:ASC,createdAt:DESC',
+            'date'    => [
+                'date:Y-m-d H:i:s',
+                'date' => 'Y-m-d H:i:s',
+            ]
+        ]);
+
+        $this->assertFalse($validation->passes());
+        $this->assertArrayHasKey('date', $validation->getInvalidData());
     }
 }